Add Houdini save lambda to unit tests and fix interval bug
Introduced a lambda function for saving simulation data to Houdini in UnitTest5_4b and UnitTest0, enabling easier logging of link positions and orientations. Also fixed a variable name in SaveDataToHoudini to use the correct interval for targetTime calculation. Updated debugger command arguments in the project user file to use a different example and parameters.